+++ client/src/components/post/CreateSubComments.js
@@ -35,7 +35,7 @@ const CreateSubComment = ({ modalVisible, setModalVisible, setSubComments, comme
   const [text, setText] = useState('');
 
   useEffect(() => {
-    axios.defaults.baseURL = 'http://10.0.2.2:3000/';
+    axios.defaults.baseURL = 'http://3.35.45.177:3000/';
     axios.defaults.headers.common['x-access-token'] = jwtToken;
     axios.defaults.headers.post['Content-Type'] = 'application/x-www-form-urlencoded';
   }, []);
