+++ client/src/utils/common.js
@@ -1,3 +1,5 @@
+import axios from 'axios';
+
 export const validateEmail = (email) => {
   const regex = /^[0-9?A-z0-9?]+(\.)?[0-9?A-z0-9?]+@[0-9?A-z]+\.[A-z]{2}.?[A-z]{0,3}$/;
   return regex.test(email);
@@ -16,4 +18,24 @@ export const validateName = (name) => {
 export const removeWhitespace = (text) => {
   const regex = /\s/g;
   return text.replace(regex, '');
+};
+
+//좌표 값으로 지명 위치 가져오기
+export const coord2address = async (location) => {
+  const { longitude, latitude } = location;
+  await axios
+    .get('https://dapi.kakao.com/v2/local/geo/coord2address.json', {
+      headers: { Authorization: 'KakaoAK 49044b35e7697c7ed1c41e61f9482595' },
+      params: {
+        x: longitude,
+        y: latitude,
+      },
+    })
+    .then((res) => {
+      const temp = JSON.parse(res.request._response)['documents'];
+      console.log('KAKAO', temp);
+
+      return temp;
+    })
+    .catch((err) => console.log('udonPeople error : ' + err));
 };
